feat: after client info to payment
